requirements.txt

python-telegram-bot==13.15
gspread
google-auth
